# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/members": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "members"
                ],
                "summary": "Create a member",
                "parameters": [
                    {
                        "description": "MemberCreateRequestDTO",
                        "name": "MemberCreateRequestDTO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.MemberCreateRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/icomphub-api_dtos.Response-dtos_MemberDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/icomphub-api_dtos.Response-any"
                        }
                    }
                }
            }
        },
        "/members/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "members"
                ],
                "summary": "Find a member",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Member ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/icomphub-api_dtos.Response-dtos_MemberDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/icomphub-api_dtos.Response-any"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "members"
                ],
                "summary": "Update a member",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Member ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "MemberUpdateRequestDTO",
                        "name": "MemberUpdateRequestDTO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.MemberUpdateRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/icomphub-api_dtos.Response-dtos_MemberDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/icomphub-api_dtos.Response-any"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "members"
                ],
                "summary": "Delete a member",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Member ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/icomphub-api_dtos.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/icomphub-api_dtos.Response-any"
                        }
                    }
                }
            }
        },
...
        "/roles": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roles"
                ],
                "summary": "Create a role",
                "parameters": [
                    {
                        "description": "RoleCreateRequestDTO",
                        "name": "RoleCreateRequestDTO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.RoleCreateRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/icomphub-api_dtos.Response-dtos_RoleDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/icomphub-api_dtos.Response-any"
                        }
                    }
                }
            }
        },
        "/roles/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roles"
                ],
                "summary": "Find a role",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/icomphub-api_dtos.Response-dtos_RoleDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/icomphub-api_dtos.Response-any"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roles"
                ],
                "summary": "Update a role",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "RoleUpdateRequestDTO",
                        "name": "RoleUpdateRequestDTO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.RoleUpdateRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/icomphub-api_dtos.Response-dtos_RoleDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/icomphub-api_dtos.Response-any"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roles"
                ],
                "summary": "Delete a role",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/icomphub-api_dtos.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/icomphub-api_dtos.Response-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "codes.Code": {
            "type": "string",
            "enum": [
                "unknow_error",
                "invalid_params",
                "get_all_users",
                "count_all_users",
                "find_user",
                "update_user",
                "create_user",
                "delete_user",
                "error_getting_users",
                "error_counting_users",
                "error_finding_user",
                "error_updating_user",
                "error_creating_user",
                "error_deleting_user",
                "get_all_technologies",
                "count_all_technologies",
                "find_technology",
                "update_technology",
                "create_technology",
                "delete_technology",
                "error_getting_technologies",
                "error_counting_technologies",
                "error_finding_technology",
                "error_updating_technology",
                "error_creating_technology",
                "error_deleting_technology",
                "get_all_projects",
                "count_all_projects",
                "find_project",
                "create_project",
                "update_project",
                "delete_project",
                "error_getting_projects",
                "error_counting_projects",
                "error_finding_project",
                "error_creating_project",
                "error_updating_project",
                "error_deleting_project",
                "get_all_class_groups",
                "count_all_class_groups",
                "find_class_group",
                "create_class_group",
                "update_class_group",
                "delete_class_group",
                "error_getting_class_groups",
                "error_counting_class_groups",
                "error_finding_class_group",
                "error_creating_class_group",
                "error_updating_class_group",
                "error_deleting_class_group",
                "login_failed",
                "login_success",
                "auth_missing_token",
                "auth_expired_token",
                "auth_invalid_token",
                "auth_valid_token",
                "role_access_denied",
                "role_insufficient",
                "file_invalid_type",
                "file_could_not_open",
                "file_could_not_find_target_dir",
                "file_could_not_save",
                "file_saved",
                "file_deleted",
                "file_found",
                "file_not_found",
                "file_could_not_delete",
                "get_all_members",
                "count_all_members",
                "find_member",
                "create_member",
                "update_member",
                "delete_member",
                "error_getting_members",
                "error_counting_members",
                "error_finding_member",
                "error_creating_member",
                "error_updating_member",
                "error_deleting_member",
                "get_all_roles",
                "count_all_roles",
                "find_role",
                "create_role",
                "update_role",
                "delete_role",
                "error_getting_roles",
                "error_counting_roles",
                "error_finding_role",
                "error_creating_role",
                "error_updating_role",
                "error_deleting_role"
            ],
            "x-enum-varnames": [
                "UnknowError",
                "InvalidParams",
                "GetAllUsers",
                "CountAllUsers",
                "FindUser",
                "UpdateUser",
                "CreateUser",
                "DeleteUser",
                "ErrorGettingAllUsers",
                "ErrorCoutingAllUsers",
                "ErrorFindingUser",
                "ErrorUpdatingUser",
                "ErrorCreatingUser",
                "ErrorDeletingUser",
                "GetAllTechnologies",
                "CountAllTechnologies",
                "FindTechnology",
                "UpdateTechnology",
                "CreateTechnology",
                "DeleteTechnology",
                "ErrorGettingAllTechnologies",
                "ErrorCoutingAllTechnologies",
                "ErrorFindingTechnology",
                "ErrorUpdatingTechnology",
                "ErrorCreatingTechnology",
                "ErrorDeletingTechnology",
                "GetAllProjects",
                "CountAllProjects",
                "FindProject",
                "CreateProject",
                "UpdateProject",
                "DeleteProject",
                "ErrorGettingAllProjects",
                "ErrorCountingAllProjects",
                "ErrorFindingProject",
                "ErrorCreatingProject",
                "ErrorUpdatingProject",
                "ErrorDeletingProject",
                "GetAllClassGroups",
                "CountAllClassGroups",
                "FindClassGroup",
                "CreateClassGroup",
                "UpdateClassGroup",
                "DeleteClassGroup",
                "ErrorGettingAllClassGroups",
                "ErrorCountingAllClassGroups",
                "ErrorFindingClassGroup",
                "ErrorCreatingClassGroup",
                "ErrorUpdatingClassGroup",
                "ErrorDeletingClassGroup",
                "LoginFailed",
                "LoginSuccess",
                "AuthMissingToken",
                "AuthExpiredToken",
                "AuthInvalidToken",
                "AuthValidToken",
                "RoleAccessDenied",
                "RoleInsufficient",
                "FileInvalidType",
                "FileCouldNotOpen",
                "FileCouldNotFindTargetDir",
                "FileCouldNotSave",
                "FileSaved",
                "FileDeleted",
                "FileFound",
                "FileNotFound",
                "FileCouldNotDelete",
                "GetAllMembers",
                "CountAllMembers",
                "FindMember",
                "CreateMember",
                "UpdateMember",
                "DeleteMember",
                "ErrorGettingAllMembers",
                "ErrorCountingAllMembers",
                "ErrorFindingMember",
                "ErrorCreatingMember",
                "ErrorUpdatingMember",
                "ErrorDeletingMember",
                "GetAllRoles",
                "CountAllRoles",
                "FindRole",
                "CreateRole",
                "UpdateRole",
                "DeleteRole",
                "ErrorGettingAllRoles",
                "ErrorCountingAllRoles",
                "ErrorFindingRole",
                "ErrorCreatingRole",
                "ErrorUpdatingRole",
                "ErrorDeletingRole"
            ]
        },
...
        "dtos.MemberCreateRequestDTO": {
            "type": "object",
            "required": [
                "nickname",
                "project_id",
                "roles"
            ],
            "properties": {
                "nickname": {
                    "type": "string",
                    "minLength": 2
                },
                "project_id": {
                    "type": "integer"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.RoleCreateRequestDTO"
                    }
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "dtos.MemberDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "project": {
                    "$ref": "#/definitions/dtos.ProjectShortDTO"
                },
                "project_id": {
                    "type": "integer"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.RoleDTO"
                    }
                },
                "status": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/dtos.UserDTO"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "dtos.MemberUpdateRequestDTO": {
            "type": "object",
            "required": [
                "roles"
            ],
            "properties": {
                "nickname": {
                    "type": "string"
                },
                "project_id": {
                    "type": "integer"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.RoleCreateRequestDTO"
                    }
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "dtos.ProjectShortDTO": {
            "type": "object",
            "properties": {
                "class_group_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "thumbnail_id": {
                    "type": "string"
                }
            }
        },
...
        "dtos.RoleCreateRequestDTO": {
            "type": "object",
            "required": [
                "name",
                "slug"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "minLength": 2
                },
                "slug": {
                    "type": "string",
                    "minLength": 2
                }
            }
        },
        "dtos.RoleDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "dtos.RoleUpdateRequestDTO": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
...
        "dtos.UserDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "personal_email": {
                    "type": "string"
                },
                "profile_picture_id": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/enums.SystemRoleEnum"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
...
        "enums.SystemRoleEnum": {
            "type": "string",
            "enum": [
                "admin",
                "user",
                "professor"
            ],
            "x-enum-varnames": [
                "RoleAdmin",
                "RoleUser",
                "RoleProfessor"
            ]
        },
...
        "icomphub-api_dtos.Response-any": {
            "type": "object",
            "properties": {
                "code": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/codes.Code"
                        }
                    ],
                    "example": "USER_CREATED"
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "User created successfully"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "icomphub-api_dtos.Response-dtos_MemberDTO": {
            "type": "object",
            "properties": {
                "code": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/codes.Code"
                        }
                    ],
                    "example": "USER_CREATED"
                },
                "data": {
                    "$ref": "#/definitions/dtos.MemberDTO"
                },
                "message": {
                    "type": "string",
                    "example": "User created successfully"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "icomphub-api_dtos.Response-dtos_RoleDTO": {
            "type": "object",
            "properties": {
                "code": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/codes.Code"
                        }
                    ],
                    "example": "USER_CREATED"
                },
                "data": {
                    "$ref": "#/definitions/dtos.RoleDTO"
                },
                "message": {
                    "type": "string",
                    "example": "User created successfully"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}